+++ src/main/groovy/org/codenarc/report/HtmlReportWriter.groovy
@@ -209,7 +209,6 @@ class HtmlReportWriter extends AbstractHtmlReportWriter {
                                 }
                             }
                         }
-
                     }
                 }
             }