+++ src/main/groovy/org/codenarc/rule/convention/InvertedIfElseRule.groovy
@@ -47,7 +47,6 @@ class InvertedIfElseAstVisitor extends AbstractAstVisitor {
         }
 
         dispatchToIfWithoutSuper ifElse
-
     }
 
     private void dispatchToIfWithoutSuper(IfStatement ifElse) {